Fix formater bug, create ElasticIntervalQuerySpec
