+++ modules/library/src/main/scala/zio/elasticsearch/ElasticQuery.scala
@@ -1,12 +1,12 @@
 /*
  * Copyright 2022 LambdaWorks
- *
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,12 +16,11 @@
 
 package zio.elasticsearch
 
-import zio.{Chunk, NonEmptyChunk}
+import zio.Chunk
 import zio.elasticsearch.ElasticPrimitive.ElasticPrimitive
 import zio.elasticsearch.data.GeoPoint
 import zio.elasticsearch.query._
 import zio.elasticsearch.script.Script
-import zio.json.ast.Json
 import zio.schema.Schema
 
 object ElasticQuery {
@@ -548,7 +547,6 @@ object ElasticQuery {
    */
   final def ids(value: String, values: String*): IdsQuery[Any] =
     Ids(values = Chunk.fromIterable(value +: values))
-
 
   /**
    * Constructs an intervals query by combining a field and an interval query.
