+++ modules/library/src/test/scala/zio/elasticsearch/ElasticQuerySpec.scala
@@ -1,12 +1,15 @@
 /*
  * Copyright 2022 LambdaWorks
  *
+
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
+
  *     http://www.apache.org/licenses/LICENSE-2.0
  *
+
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -1081,290 +1084,6 @@ object ElasticQuerySpec extends ZIOSpecDefault {
               )
             )
           )
-        },
-        test("intervalsMatchQuery") {
-          val intervalNoOptions: IntervalMatch[String] = intervalMatch("lambda works")
-
-          val intervalWithOptions: IntervalMatch[String] = intervalMatch("lambda works")
-            .withOrdered(true)
-            .withMaxGaps(2)
-            .withAnalyzer("standard")
-
-          val filter = IntervalFilter[String](
-            before = Some(intervalMatch("before_term")),
-            after = Some(intervalMatch("after_term"))
-          )
-
-          val intervalWithFilter: IntervalMatch[String] = intervalMatch("lambda works")
-            .withOrdered(true)
-            .withMaxGaps(2)
-            .withAnalyzer("standard")
-            .withFilter(filter)
-
-          val queryWithStringField = intervals("stringField", intervalWithOptions)
-          val queryWithTypedField  = intervals(TestDocument.stringField, intervalWithOptions)
-          val queryWithFilter      = intervals("stringField", intervalWithFilter)
-
-          val expectedNoOptions =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "match": {
-              |        "query": "lambda works"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedWithOptions =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "match": {
-              |        "query": "lambda works",
-              |        "analyzer": "standard",
-              |        "max_gaps": 2,
-              |        "ordered": true
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedWithFilter =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "match": {
-              |        "query": "lambda works",
-              |        "analyzer": "standard",
-              |        "max_gaps": 2,
-              |        "ordered": true,
-              |        "filter": {
-              |          "before": {
-              |            "match": {
-              |              "query": "before_term"
-              |            }
-              |          },
-              |          "after": {
-              |            "match": {
-              |              "query": "after_term"
-              |            }
-              |          }
-              |        }
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          assert(intervalNoOptions)(
-            equalTo(
-              IntervalMatch[String](
-                query = "lambda works",
-                analyzer = None,
-                useField = None,
-                maxGaps = None,
-                ordered = None,
-                filter = None
-              )
-            )
-          ) &&
-          assert(intervals("stringField", intervalNoOptions).toJson(None))(
-            equalTo(expectedNoOptions.toJson)
-          ) &&
-          assert(queryWithStringField.toJson(None))(
-            equalTo(expectedWithOptions.toJson)
-          ) &&
-          assert(queryWithTypedField.toJson(None))(
-            equalTo(expectedWithOptions.toJson)
-          ) &&
-          assert(queryWithFilter.toJson(None))(
-            equalTo(expectedWithFilter.toJson)
-          )
-        },
-        test("intervalRange") {
-          val intervalNoBounds = intervalRange(
-            lower = None,
-            upper = None,
-            analyzer = None,
-            useField = None
-          )
-
-          val intervalWithBounds = intervalRange(
-            lower = Some(Left("10")),
-            upper = Some(Right("20")),
-            analyzer = Some("standard"),
-            useField = Some("otherField")
-          )
-
-          val intervalWithOnlyGt = intervalRange(
-            lower = Some(Left("10")),
-            upper = None,
-            analyzer = Some("standard"),
-            useField = Some("otherField")
-          )
-
-          val queryWithStringField = intervals("stringField", intervalWithBounds)
-          val queryWithTypedField  = intervals(TestDocument.stringField, intervalWithBounds)
-          val queryWithOnlyGt      = intervals("stringField", intervalWithOnlyGt)
-
-          val expectedNoBounds =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "range": {}
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedWithBounds =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "range": {
-              |        "gt": "10",
-              |        "lte": "20",
-              |        "analyzer": "standard",
-              |        "use_field": "otherField"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedWithOnlyGt =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "range": {
-              |        "gt": "10",
-              |        "analyzer": "standard",
-              |        "use_field": "otherField"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          assert(intervalNoBounds)(
-            equalTo(
-              IntervalRange(
-                lower = None,
-                upper = None,
-                analyzer = None,
-                useField = None
-              )
-            )
-          ) &&
-          assert(queryWithStringField.toJson(None))(
-            equalTo(expectedWithBounds.toJson)
-          ) &&
-          assert(queryWithTypedField.toJson(None))(
-            equalTo(expectedWithBounds.toJson)
-          ) &&
-          assert(queryWithOnlyGt.toJson(None))(
-            equalTo(expectedWithOnlyGt.toJson)
-          )
-        },
-        test("intervalWildcard") {
-          val wildcardExact: IntervalWildcard[String] =
-            intervalWildcard("la*mb?da")
-
-          val wildcardContains: IntervalWildcard[String] =
-            intervalContains("lambda")
-
-          val wildcardStartsWith: IntervalWildcard[String] =
-            intervalStartsWith("lambda")
-
-          val wildcardEndsWith: IntervalWildcard[String] =
-            intervalEndsWith("lambda")
-
-          val queryExact: Intervals[String] =
-            Intervals("stringField", wildcardExact)
-
-          val queryContains: Intervals[String] =
-            Intervals("stringField", wildcardContains)
-
-          val queryStartsWith: Intervals[String] =
-            Intervals("stringField", wildcardStartsWith)
-
-          val queryEndsWith: Intervals[String] =
-            Intervals("stringField", wildcardEndsWith)
-
-          val expectedExact =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "wildcard": {
-              |        "pattern": "la*mb?da"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedContains =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "wildcard": {
-              |        "pattern": "*lambda*"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedStartsWith =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "wildcard": {
-              |        "pattern": "lambda*"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          val expectedEndsWith =
-            """
-              |{
-              |  "intervals": {
-              |    "stringField": {
-              |      "wildcard": {
-              |        "pattern": "*lambda"
-              |      }
-              |    }
-              |  }
-              |}
-              |""".stripMargin
-
-          assert(wildcardExact)(
-            equalTo(
-              IntervalWildcard[String](
-                pattern = "la*mb?da",
-                analyzer = None,
-                useField = None
-              )
-            )
-          ) &&
-          assert(queryExact.toJson(None))(equalTo(expectedExact.toJson)) &&
-          assert(queryContains.toJson(None))(equalTo(expectedContains.toJson)) &&
-          assert(queryStartsWith.toJson(None))(equalTo(expectedStartsWith.toJson)) &&
-          assert(queryEndsWith.toJson(None))(equalTo(expectedEndsWith.toJson))
         },
         test("kNN") {
           val queryString         = kNN("stringField", 5, 10, Chunk(1.1, 2.2, 3.3))
